+++ style_bert_vits2/nlp/chinese/opencpop-strict.txt
